Compare team stack starting from the current index instead of the whole stack Use index from stack to get team during method invocation instead of binding it in handleReplace / handleBefore; this also requires correct index incrementation while executing the DAG Adapt handleReplace for callins with variable argument lengths Simplify handleReplace / handleBefore by swapping arguments of lift MethodHandle instead of liftToRoleCollect MethodHandle
Not comparing the whole team stack allows cached invocations to  be reused more often, since only teams following the current point of execution have to be equal to the cached stack. The second point is required for this to work, as before the index might have had an incorrect value when invoking a cached handle with a different team stack.